requirements.txt
flask
flask-sqlalchemy
python-dotenv
psycopg2-binary